pyproject.toml
[project]
name = "mtd"
version = "0.1.0"
description = "Yet another Markdown to Documents converter, lightweight and themeable."
readme = "README.md"
requires-python = ">=3.14"
license = "MIT"
dependencies = [
    "click>=8.1",
    "markdown>=3.5",
    "python-docx>=1.1",
    "odfpy>=1.4",
    "pyyaml>=6.0",
    "jinja2>=3.1",
]

[project.scripts]
mtd = "mtd.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ruff>=0.9",
    "pytest>=8.0",
    "mypy>=1.15",
    "pre-commit>=4.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "N",      # pep8-naming
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "RUF",    # ruff-specific
]

[tool.ruff.lint.isort]
known-first-party = ["mtd"]

[tool.mypy]
python_version = "3.14"
strict = false
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
check_untyped_defs = true
